backend/src/requirements.txt
appdirs==1.4.3
attrs==18.1.0
black==18.6b4
bokeh==0.13.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.0
oauthlib==2.1.0
packaging==17.1
pipenv==2018.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pygame==1.9.3
PyJWT==1.6.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
toml==0.9.4
tornado==5.1
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==4.0
